Asynchronicity event rewrite
This is a full rewrite of how event handling works for Daemon interaction.
The new design flow schedules execution to the Daemon on their own thread,
waiting to acquire a lock through an MVar primitive. This ensures that
interaction with the Ghcid handle is done synchronously, while still
allowing interrupts.

We're making use of Brick's Bounded Channels, which allow asynchronous
message passing to the Brick application itself. Messaging passing must
now be done through the CustomAppEvent.

This is feature complete but there's a lot of clean up that still
must be done before this should land.