Force upgrade fsnotify
This breaks for whatever reason only on hackage. It appears
ghcid depends on fsnotify as an executable, but not as a
library. We're not using it as an executable, but perhaps
the documentation is breaking on it. I'm uncertain why
we're building the executable dependencies, but until that's
fixed we can't do anything except either downgrade text
or force upgrade fsnotify.